Changes to guess_optimization
Every application should normalize its IPS history to calculate the efficiency score relative to its own fair share performance.
Need to do that for QoS to have any meaning and for giving applications opportunity to cooperate.

Also, there is a lot of oscillatory behavior. Need to either build some buffering or identify the oscillation
through a success measure of a particular shift in number of contexts. Need to gather more data on this.