Two changes to budgeting numbers.
- per_app_cpu_budget in [SAM_MIN_CONTEXTS, cpuinfo->total_cpus]
- change how needs_more is computed by changing initial_remaining_cpus.
This means that an application will have to take CPUs from other
applications if its request is greater than what it currently has (and
no other application has given up contexts). The point of this is so
that the first applications don't assume everything is free, and now
instead they'll have an appropriate needs_more count.